spanner-pub-sun-cr/requirements.txt
Flask
gunicorn
google-cloud-storage
google-cloud-spanner
google-cloud-pubsub